misc service method renames and add waste_services for retrieving waste information
