+++ server/apps/trak/services/manifest_services.py
